new separate files for tests and userstory10
